requirements.txt
Flask==2.3.1
google-generativeai==0.3.1
gunicorn==21.2.0
Pillow==9.4.0
python-docx==1.1.0

